pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecommerce-retention-project"
version = "0.1.0"
description = "End-to-end ecommerce retention modeling project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2",
  "numpy>=1.26",
  "scikit-learn>=1.5",
  "pyyaml>=6.0",
  "joblib>=1.4",
  "matplotlib>=3.9",
  "seaborn>=0.13",
  "fastapi>=0.115",
  "uvicorn>=0.32",
  "pydantic>=2.9",
  "pyarrow>=17.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "ruff>=0.6",
  "black>=24.8",
  "pre-commit>=3.8"
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
